requirements.txt
Flask>=2.0
opencv-python
pytesseract
Pillow
deep-translator
werkzeug
